+++ src/extension.ts
@@ -5,7 +5,7 @@ import * as vscode from "vscode";
 import { MessageService } from "./services/message.service";
 import { UserService } from "./services/user.service";
 import { MainTreeContainer } from "./views/tree/main.tree";
-import { TaskWebview } from "./views/webview/main.webview";
+import { TaskWebview } from "./views/webview/task.webview";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -23,7 +23,6 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // Set Api Key through Axios interceptor
   const apiKey = await vscode.workspace.getConfiguration("hacknplan").get("apiKey");
-  console.log(apiKey);                                                           
   if (apiKey) {
     Axios.interceptors.request.use((config: AxiosRequestConfig) => {
       config.headers.Authorization = `ApiKey ${apiKey}`;
@@ -38,81 +37,11 @@ export async function activate(context: vscode.ExtensionContext) {
 
   // const mainTree = new MainTreeContainer();
 
-  const taskWebview = new TaskWebview(context.extensionPath);
+  const path = context.extensionPath;
+
+  const taskWebview = new TaskWebview(path);
 
   context.subscriptions.push(taskWebview.showTaskCommand);
-
-
-    // const showTaskDetailCommand = vscode.commands.registerCommand(
-    //   "hacknplan.showTask",
-    //   async (task: Task) => {
-    //     const panel = vscode.window.createWebviewPanel(
-    //       "showTaskDetail",
-    //       "Show Task Detail",
-    //       vscode.ViewColumn.Beside,
-    //       { enableScripts: true }
-    //     );
-
-    //     const showTaskCSS = vscode.Uri.file(
-    //       path.join(
-    //         context.extensionPath,
-    //         "src",
-    //         "views",
-    //         "webview",
-    //         "showTask",
-    //         "styles.css"
-    //       )
-    //     );
-    //     const showTaskCSSSrc = panel.webview.asWebviewUri(showTaskCSS);
-
-    //     const showTaskJS = vscode.Uri.file(
-    //       path.join(
-    //         context.extensionPath,
-    //         "src",
-    //         "views",
-    //         "webview",
-    //         "showTask",
-    //         "utils.js"
-    //       )
-    //     );
-    //     const showTaskJSSrc = panel.webview.asWebviewUri(showTaskJS);
-
-    //     panel.webview.html = await ShowTask.showTaskHTML(
-    //       task,
-    //       showTaskCSSSrc,
-    //       showTaskJSSrc
-    //     );
-
-    //     // handle patch of task
-    //     panel.webview.onDidReceiveMessage(async (message: IMessageTask) => {
-    //       switch (message.command) {
-    //         case "createNewSubTask":
-    //           const newSubTask = await TaskService.createNewSubTask(
-    //             message.params
-    //           );
-    //           return panel.webview.postMessage({
-    //             command: "createNewSubTaskResponse",
-    //             params: {
-    //               newSubTask,
-    //             },
-    //           });
-    //         case "addAssignedUser":
-    //           await TaskService.addUserToTask(message.params);
-    //           return MainTreeContainer.taskTreeProvider.refresh();
-    //         case "deleteAssignedUser":
-    //           await TaskService.deleteUserFromTask(message.params);
-    //           return MainTreeContainer.taskTreeProvider.refresh();
-    //         case "updateSubTask":
-    //           return await TaskService.updateSubTask(message.params);
-    //         case "deleteSubTask":
-    //           return await TaskService.deleteSubTask(message.params);
-    //         case "updateTask":
-    //           await TaskService.updateTask(message.params);
-    //           return MainTreeContainer.taskTreeProvider.refresh();
-    //       }
-    //     });
-    //   }
-    // );
 
 }
