+++ src/models/stage.model.ts
@@ -22,7 +22,6 @@ export class StageTreeItem extends vscode.TreeItem {
   constructor(
     public readonly label: string,
     public readonly collapsibleState: vscode.TreeItemCollapsibleState,
-    public readonly type: ModelType,
     public readonly boardId: number,
     public readonly projectId: number,
     public readonly stageId: number
